+++ crates/cli/src/commands/docker/prune.rs
@@ -45,15 +45,15 @@ pub async fn prune_node(
         .await?;
 
     // Remove extraneous node module folders for unfocused projects
-    let mut futures = vec![];
+    // let mut futures = vec![];
 
-    for project_id in &manifest.unfocused_projects {
-        if let Some(project_source) = workspace.projects.projects_map.get(project_id) {
-            futures.push(fs::remove_dir_all(
-                workspace.root.join(project_source).join(NODE.vendor_dir),
-            ));
-        }
-    }
+    // for project_id in &manifest.unfocused_projects {
+    //     if let Some(project_source) = workspace.projects.projects_map.get(project_id) {
+    //         futures.push(fs::remove_dir_all(
+    //             workspace.root.join(project_source).join(NODE.vendor_dir),
+    //         ));
+    //     }
+    // }
 
     try_join_all(futures).await?;