+++ packages/cli/CHANGELOG.md
@@ -43,6 +43,7 @@
 - When writing YAML files, indentation and formatting will be inferred from the closest
   `.editorconfig` as best as possible.
 - Fixed an issue where parsing `yarn.lock` would panic on certain Windows machines.
+- Fixed an issue where `moon docker prune` would remove required node modules.
 
 #### ⚙️ Internal